// src/app/model/IItem.ts
export interface IItem {
    id: number;
    name: string;
    price: number;
    sale: number;
    color: string;
    size: string;
    length: string;
    url: string;
    quantity: number;
    images: string[];
    state: number;
    details: string;
    materialAndCare: string;
    product_type_id: number;
}